Start again, better this time.

+++ property_file_reader.rb
@@ -1,5 +1,3 @@
-# Not independently executable
-
 =begin
 --------------------------------------------------------------------------------
 
@@ -8,6 +6,24 @@ properties.
 
 --------------------------------------------------------------------------------
 =end
+
+class FunkyHash
+  def self.create()
+    hash = {}
+
+    def hash.method_missing method_id, *args
+      if args.empty?
+        self[method_id.to_s]
+      elsif method_id.to_s.end_with?('=')
+        self[method_id.to_s.chop] = args[0]
+      else
+        super
+      end
+    end
+
+    hash
+  end
+end
 
 class PropertyFileReader
   # Read a properties file and return a hash.
@@ -18,20 +34,22 @@ class PropertyFileReader
   # the path to the properties file.
   #
   def self.read(file_path)
-    properties = {}
+    properties = FunkyHash.create
     properties["properties_file_path"] = File.expand_path(file_path)
 
-    File.open(file_path) do |file|
-      file.each_line do |line|
-        line.strip!
-        if line.length == 0 || line[0] == ?# || line[0] == ?!
-          # ignore blank lines, and lines starting with '#' or '!'.
-        elsif line =~ /(.*?)\s*[=:]\s*(.*)/
-          # key and value are separated by '=' or ':' and optional whitespace.
-          properties[$1.strip] = $2
-        else
-          # No '=' or ':' means that the value is empty.
-          properties[line] = ''
+    if File.exist?(file_path)
+      File.open(file_path) do |file|
+        file.each_line do |line|
+          line.strip!
+          if line.length == 0 || line[0] == ?# || line[0] == ?!
+            # ignore blank lines, and lines starting with '#' or '!'.
+          elsif line =~ /(.*?)\s*[=:]\s*(.*)/
+            # key and value are separated by '=' or ':' and optional whitespace.
+            properties[$1.strip] = $2
+          else
+            # No '=' or ':' means that the value is empty.
+            properties[line] = ''
+          end
         end
       end
     end
